Add CLI for fix_preserve_terms and share duration formatter.
Extract format_remaining_duration_zh into gui_qt/duration_format.py and replace fix_preserve_terms hardcoded paths with manual/results subcommands.
