+++ gui_qt/workflow_progress.py
@@ -5,6 +5,8 @@ import re
 import time
 from dataclasses import dataclass, replace
 
+from .duration_format import format_remaining_duration_zh
+
 
 @dataclass(frozen=True)
 class WorkflowProgressState:
@@ -454,28 +456,6 @@ def _with_progress_timing(
         seconds_per_unit=seconds_per_unit,
         remaining_seconds=remaining_seconds,
     )
-
-
-def format_remaining_duration_zh(seconds: int) -> str:
-    if seconds <= 0:
-        return "即将完成"
-    if seconds < 60:
-        return f"约剩 {seconds} 秒"
-    minutes, secs = divmod(seconds, 60)
-    if seconds < 3600:
-        if secs >= 30:
-            minutes += 1
-            secs = 0
-        if secs:
-            return f"约剩 {minutes} 分 {secs} 秒"
-        return f"约剩 {minutes} 分"
-    hours, minutes = divmod(minutes, 60)
-    if minutes >= 30:
-        hours += 1
-        minutes = 0
-    if minutes:
-        return f"约剩 {hours} 小时 {minutes} 分"
-    return f"约剩 {hours} 小时"
 
 
 def _current_file_index(state: WorkflowProgressState) -> int:
